deep-research-bot/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# HTTP & Scraping
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
newspaper3k==0.2.8
lxml==4.9.3

# AI
openai==1.3.7

# YouTube
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
rich==13.7.0

